Import auth_token middleware from keystoneclient
Although the master auth_token file is now in keystoneclient, it will take
some time to get all the paste files to point to it there rather than here.
Hence, we import it back here to provide backward compatibility for a release
or so, after which we will remove it from the server.

Change-Id: Iccdb7839a611cdda233e4ea96f68c64d6d82f49c