Fix middleware logging for swift
Swift sets 'log_name' in the conf, and this can be used to set the
correct logger.  This patch moves the log configuration to
initialization so that the logger can be set appropriately and the
middleware will log to the swift proxy logs.  This also requires the
following review in swift to work:
https://review.openstack.org/#/c/17983/

Fixes bug #1089664

Change-Id: I315102ba277ac1f5ff92f4d79ae745bf53559328