Make auth_token lazy load the auth_version.
Updates recent changes to the auth_token middleware (this
is a regression in d782a99) so that self.auth_version is
lazy loaded.

This fixes issues where other openstack services would fail
to startup correctly if Keystone is not running. The issue
was auth_token was trying to make a request to '/' to get
versions information on startup to "autodetect" the correct
version to use.

This patch fixes startup issues by moving the version detection
so that it is lazy loaded right before it is actually used.

This issue should fix SmokeStack :)

Fixes LP Bug #1154806.

Change-Id: Ib24f5386fa1ffe0e0365548840f0cfeaae36f548