Copy s3_token middleware from keystone
The s3_token middleware was in keystone but it should be in
python-keystoneclient. As a first step to removing s3_token from
keystone, copy it to python-keystoneclient. Also copy the tests.

A couple of changes were required

- Changed oslo-incubator imports from keystone to keystoneclient.
- Changed logging from oslo-incubator to use Python logging and
  configured the same way as the auth_token middleware.

I checked and I didn't see anything to be added to requirements.txt
or test-requirements.txt.

bp s3-token-to-keystoneclient

Change-Id: I64fef4101180e5aa661442d538c3237bdad0c37c
Closes-Bug: #1178741