Add service catalog to domain scoped token fixture
This was simply forgotten previously and testing that required a service
catalog was done with the project token. This has led to some problems
of not testing the domain paths correctly.

Fix the one test that was incorrectly relying on an empty service
catalog being returned from a domain scoped token.

Change-Id: I66e3a9302d6a73ba17d4ed45f2366d9bb91461ab