Fix auth_token middleware test invalid cross-device link issue
Fixes path of temporary files, containing token revocation lists,
which weren't in the appropriate directory. File renaming now works
properly since source and destination files are in the same directory,
and therefore on the same partition.

Added to .gitignore temporary files generated during tests, that are
not being removed at the end of test.

Change-Id: I21ff81419313f4205ebe552556e27b86135fd096
Closes-Bug: #1265190