Move auth_token cache pool tests out of NoMemcache
There were cache pool tests in NoMemcacheAuthToken. Since this class
un-imports memcache, setting the memcached_servers option would just
be ignored anyways.

The tests were moved into a separate class and changed to not set
memcached_servers. The behavior should be the same.

Also, there was a cache pool test in CommonAuthTokenMiddlewareTest
that fit better in this new class.

Change-Id: Icfce7398f2a29753200c2010bba58924980622c0