Expose an entry point to list auth_token middleware config options
Register a 'keystonemiddleware.auth_token' entry point in the
oslo.config.opts namespace which, when called, returns a list of the
configuration options which may be registered by the project at runtime.

The idea here is that the sample config file generator can query this
and include the returned options in the sample config file of any
applications which use the middleware, e.g. currently the options
were listed in 'keystone_authtoken' section for those major projects.

Related-Change-Id: I15686708fc9460948a58cfea3d18dae40ba1fda9
Related-Change-Id: Ic28351258652d2ea38974e2f4d1aa6b1d3dd7192
Related-Change-Id: I76043b08e2872867e5af2a5ac902e4d092fda5c8
Related-Bug: #1300546

Change-Id: Iae31856d5886ee78786972d80c7c103c3460a2b3
Signed-off-by: Zhi Yan Liu <zhiyanl@cn.ibm.com>