Mark keystonemiddleware as being a universal wheel
Keystonemiddleware is pure python and is version agnostic. 
It should be marked as a universal wheel as it can be installed 
anywhere by pip (see the documentation on python packaging and
wheels). This mirrors python-keystoneclient's marking as a
universal wheel (as most of the code in keystonemiddleware
originated from python-keystoneclient).
 
Change-Id: I3e25578e6fc7a30b5eb7544d85a7ef39e711bd34