Convert auth_token middleware to use sessions
With this patch, session objects will be used for requests
and token management.

It is no longer permissable to specify both a username/password
and a admin_token. This used to work but now you get one plugin
or the other.

There is one test removed in this patch which was to do with having the
auth token refreshed if it was stale. This is no longer handled by the
middleware but expected to be managed by the auth plugin.

This fixes the existing behaviour that if an admin_token was given and
was marked invalid then the middleware would fallback to using the
username and password provided. If an authentication method fails then
this is something that should be addressed not compensated for.

Co-authored-by: Harry Rybacki <hrybacki@redhat.com>
Change-Id: Ib52beaaa1e01875cceaae78dc879a6399ccefa36
Closes-Bug: #1307252