Use connection retrying from keystoneclient
Keystoneclient added retrying requests on connection errors in 0.11. We
should use this in preference to our own scheme.

This is not exactly the same as the existing code. The existing code
catches any non-HTTP exceptions and retries the request. This has been a
problem in the past for example with AttributeErrors and such being
retried that will never succeed.

This implementation that will only retry when you get an actual
connection error is more correct.

I also remove the NetworkError exception. This is only used internally
and it is easier to simply catch the connection exceptions where the
NetworkError used to be.

Change-Id: I6c3c971ea1964cfbf50519e8faf0a0c7dff47598