Fix environ keys missing HTTP_ prefix
The _fmt_msg() function constructs a debug message using
identity attributes from the request environ dict, but
accesses these using keys such as X_USER_ID, when the
correct key should be HTTP_X_USER_ID. Consequently no
identity attribute values are ever included in the debug
log message.

This patch adds the missing HTTP_ prefix to the keys and
adds an assertion to the unit tests to verify that the debug
log is generated as expected.

Closes-Bug: 1408602
Change-Id: Ib309abcb3d0f15a20ae581105de1f8964497e5a2