Allow loading auth plugins via overrides
Swift configures auth_token purely by paste options, not oslo.config.
This means that we cannot rely on purely using the keystoneclient
load_from_config_options for auth plugins.

Copy the logic from keystoneclient regarding auth plugin loading from
config files and make it specific to the _conf_get that auth_token
middleware uses so that the auth plugin options obey options from paste
as other options do.

This will be replaced with a keystoneclient helper as soon as possible,
however we want to fix the swift issue before that happens and we will
still be compatible going forward.

Change-Id: I54ac4c566cb798196ea18e24d4ce868450f269f3
Closes-Bug: #1428900