Simplify request making in auth_token tests
One of the reasons that there are so many unrelated tests in the basic
test_auth_token_middleware tests file is that there is a lot of token
and mocking setup that is required for all tests.

Firstly extract a base class with some middleware creation helpers. Use
these in the existing tests. Remove the concept of start_request. webob
has helpers that can simplify this into the standard request/response
handling so there is no reason we need this and we don't want to copy it
to new test files.

Change-Id: I31bb40feb1915b75193eac621be5501dde162f08