Fix list_opts test to not check all deps
The list_opts entrypoint test failed unnecessarily when the
dependencies in the packages were inconsistent. This test doesn't
need to verify that the dependencies are consistent, only that the
entrypoint is available and provides the expected function.

For the backport, I needed to remove the stevedore additions to
test-requirements-py3.txt and test-requirements.txt because
stevedore is already in requirements.txt.

Change-Id: I857222e0637b09f45352d9cabbdc5cac8e4fd562
(cherry picked from commit 03d6f0ce9df169f43696ecade16ec1cf4162647f)