Don't cache signed tokens
Currently the cache is always checked and a result is always cached even
if it is for a PKI token. In the case of PKI token though we don't have
to do an online lookup, we just decoded the data we've been given.

There is a CPU vs IO bound tradeoff as to whether you take decrypt the
token or wait for memcache to return however I do not see any point in
storing the result of a signing operation.

This will also make it much easier to support similar signing
validations on the keystone server side because we are not bound by the
same cache constraints.

Change-Id: Ibb45b14954ef8a26dcc0e3b5dfa7c047dd8edfe1