Separate the fetch and validate parts of auth_token
Split the validate process into two distinct parts. The fetch process
will retrieve the token data, either from PKI or from the identity
server. It is also responsible for caching this information.

The validate process then looks at the data returned and decides whether
it is allowed for this request. A simplification for later is that token
binding is performed in addition to the validate step because this
really should be performed once per request and not once per token.

Because of this we are no longer caching the validation of a request,
only the actual request data. We therefore remove some tests that were
testing that the validation of a token was cached.

Change-Id: Icbd44f5edbeadb1a52cccf2a481a1e861b0de120