Allow to use oslo.config without global CONF
If an application doesn't use a global configuration object and load
the middleware with api paste, they are no ways to read the
configuration options from the application configuration file.

This change fixes that, the api paste config will looks like:

  [filter:authtoken]
  paste.filter_factory = keystonemiddleware.auth_token:filter_factory
  oslo_config_project = aodh

With this, the keystonemiddleware will automatically load
the configuration of the project aodh with a local oslo.config object
instead of the global one.

This allows application to not rely of the global oslo.config object
and continue to use paste and keystonemiddleware.

Closes-bug: #1482078
Related-bug: #1406218
Change-Id: I48c3d6a6a5486c9c035a15a75c025be7f5abaab4