Create Environment cache pool
We have three distinct ways that a cache pool can be created, however
only two cache pool classes with the regular _CachePool doubling up.
There's no reason for this and it's harder to read.

Create a very simple _EnvCachePool for the situation where a cache
object was passed down through environment variables.

Change-Id: I86147e887efd6ae302caa12ce996be697d36aa17