only make token invalid when it really is
The auth_token middleware was marking a token invalid in the cache
on any otherwise-uncaught exception, not only when it has really
determined that the token is invalid. This will fix that by being
better about using the InvalidToken exception only when the token is
really invalid, and only marking invalid in the cache when that
exception occurs.

This fix uncovered a bug in the auth_token unit test
test_composite_auth_delay_invalid_service_and_user_tokens, so that has
been addressed as well. It was attempting to use "invalid-user-token",
which was not initialized, where it meant to use "invalid-token". This
was leading to a NoMockAddress exception. Prior to this fix, the code
turned all unexpected exceptions into InvalidToken exceptions, masking
this issue.

It also uncovered DeprecationWarnings raised in _identity.py for usage
of logging.warn rather than logging.warning in py34. That has also
been fixed. As above, prior to this fix those would have been turned
into InvalidToken exceptions (incorrectly).

Closes-Bug: #1488267
(cherry picked from commit 7dcbdf774803a92bfc24704dec5f919ac75ad315)
(cherry picked from commit 75fbc5910afad5d6dbce5506cf9b38f793a68644)

Conflicts:
	keystonemiddleware/auth_token/__init__.py
	keystonemiddleware/auth_token/_identity.py
	keystonemiddleware/tests/auth_token/test_auth_token_middleware.py

Cherry-Pick Note: The main conflict here is that webob is not used in
this version of auth_token middleware and so you can't simply raise a
500/503 error from deep in the code and expect it to be handled.
Currently the ServiceError handles the 503 exception. I added a
_InternalServiceError exception that does the same job as the similar
webob exception.

Change-Id: I2e487fb02c9171f743ecc1f4f230b29a5e96212d