Remove auth headers in AuthProtocol
remove_auth_headers is the function that removes all the
HTTP_X_PROJECT_ID and similar authentication headers from an incoming
request so that you can't fake authentication to a service.

It turns out that we pass X-Project-Id and X-Domain-Id domain headers to
keystone for tokenless auth (and keystone doesn't use those headers
otherwise anyway) so if we want to use auth_token middleware in keystone
we only want to strip headers in the AuthProtocol subclass.

Change-Id: I5f0ea022156d0aaff4b3dc71c1cba981687723c9