Make BaseAuthProtocol public
BaseAuthProtocol is the basic validation middleware logic without all
the extra hacks, caching, CONF handling etc that has built up around
auth_token middleware. It is the basic class that should be reusable
with keystone and any newer implementations of middleware.

It was private whilst the interfaces stabilized however it has not
cheanged in recent releases and should now be available for use.

Change-Id: I1e935f5736847941b11495ae086694092baf8b3f