ignore deprecation calls in test_audit_middleware
pycadf now issues deprecation warnings on certain events. this
behaviour was causing the the stable/liberty branch to fail, since
we immediately fail any time keystonemiddleware calls a deprecated
function. for the test suite that calls pycadf functions, we should
ignore these errors. Note that this does not happen on the master
branch.

Change-Id: I76cc17b8e4341ee73f343445b3f1b99b64a4977e
Related-Bug: 1504889