Revert "Stop using private keystoneclient functions"
This reverts commit 96f6668a27c34c7fd49c8a0df160c0789b0aa6ba.

Whilst in theory I agree we don't want to expose this many positional
arguments. This is exactly what positional is supposed to be used for.

I'm happy to have the debate as to where we can make @positional
publically available from but for now (and follow up patch) let's keep
it.

Change-Id: Idd07e49370597b63c0e49b56843266a7014a0e65