Group common PKI validation code - Tests
The PKI token validation for PKI and PKIZ tokens both individually do
the cms validation and the revocation checking. There's no reason that
this needs to be in both the functions. We can split the decode process
and standardize the validation process to make the code flow easier to
understand.

Because _verify_signed_token and _verify_pkiz_token are going away in
favour of the common _validate_offline function move the tests up to
this function. This should show that there is no behaviour difference
with the coming refactor.

Change-Id: I9191c6a1957a8def1c936cd3015cfbe737142d3f