Deprecate in-process cache
For a long time now if you don't configure memcache then
auth_token middleware would cache the tokens in process
memory.

This is not the job of auth_token middleware. If you need to
cache you should configure memcache otherwise auth_token will
authenticate with keystone for every token request.

As such, this feature is deprecated and may be removed in the
5.0.0 release or the "O" development cycle (whichever is later).

Change-Id: Ied2b88c8cefe5655a88d0c2f334de04e588fa75a