Determine project name from oslo_config or local config
For services such as Swift, which may not be utilizing oslo_config, we need
to be able to determine the project name from local config. If project name
is specified in both local config and oslo_config, the one in local config will
be used instead.

In case project is undetermined (i.e. not set), we use taxonomy.UNKNOWN
as an indicator so operators can take corrective actions.

Change-Id: Ia95cbc9974d0c39c6b77d966cffdef2885350d77
Closes-Bug: 1583690