Move auth token opts calculation into auth_token
The list of all auth token opts is currently calculated in opts.py.
That module is included in auth_token/__init__.py, which in turn owns
some opts that are needed by the former.

This creates a circular import dependency.

In order to fix such situation, this patch proposes to move the auth
token opts calculation into auth_token/__init__.py, so that it will
no longer need opts.py.

Co-Authored-By: Alfredo Moralejo <amoralej@redhat.com>

Closes-Bug: #1591913
Change-Id: If67d8bdb68a5ab9c07b960ad0111e2310ad82c83