Fix an issue with oslo_config_project paste config
When oslo_config_project is specified in the paste config, we need to
lookup the configuration file properly. Commit
f8c150a9cc9b407b2df87244daf3342177260e90 changed the way we looked up
for the default config files, and passed an empty list instead of None,
which means that no look were done anymore. This restores the previous
behavior.

Change-Id: I6aec548cec24dfcd48e4c116c07f53cc097c6059
Closes-Bug: #1594930