Use AccessInfo in UserAuthPlugin instead of custom
The UserAuthPlugin in auth_token_middleware provided a sanitized view of
an AccessInfo object. This was required in keystoneclient time when it
would do things like default to the 'default' domain on v2 tokens which
was wrong and should not be exposed to other services.

This was one of the things that was cleaned up in the keystoneauth
transition and so there's really no further need to hide this from
services. It also meant every time we wanted to expose something
additional to a service we had to hack in a proxy property.

Remove the old _TokenData proxy and expose the keystoneauth AccessInfo
object to services.

Change-Id: I95a7a7a7ad59daf6cb623923de6c8a8533c4734a