Mock log only after app creation
The unit test that is getting fixed fails with error:

AttributeError: None does not have the attribute 'info'

It happens because _LOG.info is being patched at the moment when _LOG
is None. For a reason, this doesn't happen always. For example,
the problem was gone after a recheck in
https://review.openstack.org/#/c/391130/.

Patch log only after app creation. This seems to be a common pattern
across all other tests involving mock, for example, in
test_conf_middleware_log_and_default_as_messaging.

Change-Id: Ia48520d6835c5712d77e182cffbf50cc6b56b9ce