Add Constraints support
Adding constraints support to libraries is slightly more complex than
services as the libraries themselves are listed in upper-constraints.txt
which leads to errors that you can't install a specific version and a
constrained version.

This change adds constraints support by also adding a helper script to
edit the constraints to remove keystonemiddleware.

Change-Id: Ic181afad5bf3dd89a2edfdf9fb3026f85cdc37a6