Fix gate for newton
The gate is wedged and this is a combination of two commits.

1. Auth token, set the correct charset

Explicit set the charset to UTF-8 when replying with 401. It's now
required by webob.

A previous attempt to fix this problem was introduced as part of the
commit 20fb1dbe5a884ecdbf6ee5ea26b2660f7a7e4b1e but this is only
compatible with the version 1.7.0 of WebOb, older versions will break.
See:
http://logs.openstack.org/05/371605/12/check/gate-ironic-python35-db/686bd86/console.html#_2017-01-03_11_10_11_760197

Change-Id: I30adc87882ce8bd2c18588161d5c60d77a56925e
Closes-Bug: #1653646
(cherry-picked from: 2092d5783607223c7cbdfed690997e9c9f831b6f)

2. Add docutils contraint on 0.13.1 to fix building

See:
http://lists.openstack.org/pipermail/openstack-dev/2016-December/108742.html

Change-Id: Ied3c922b8bbb52a3589c2f5308ecf87fc1415fbc
(cherry-picked from: 29a879c0ed5d1af5dca0cf6fd0394487f97b1670)