Only include response body if there's a response
When handling timeouts talking to the keystone server, you may see log
messages like

  authtoken: Bad response code while validating token: 408
  authtoken: Token validation failure.
  <traceback>
  AttributeError: 'NoneType' object has no attribute 'text'

Since there's no response from the server when keystoneclient raises
RequestTimeout [1], the `response` attribute is understandably None.

Now, only log the response text if there's text to log. Additionally,
log the response message (as well as status code) to provide as much
context as we can for the error.

[1] https://github.com/openstack/python-keystoneclient/blob/3.15.0/keystoneclient/session.py#L469

Change-Id: Id400e4c38d07cbe7e1866dd572a17fc54c31e82a