Remove tox_install.sh
As part of removing reliance on the old and deprecated zuul-cloner, we
need to shift constraints declaration to the deps line. This means we
unfortunately have to duplicate the extras declarations into
test-requirements - because otherwise the contraints for
keystonemiddleware conflicts with the installation of itself.

Conflicts:
 Updated tox.ini to pull stable/rocky constraints

Change-Id: I8dbb31d1c1fda6df386f456dcf1d8bbed6d168ce
(cherry picked from commit 9a6875d479c22b81b47f545253f674a8c6221331)