Documentation Fix - auth_url Port Number
Made a small fix to the documentation - replacing
the current auth_url port number 35357, in the
configuration section of the [keystone_authtoken],
with 5000.

This was based on an online conversation with Colleen;
with the removal of the v2 API from keystone the project
now recommends use of port 5000 instead of the previous one.

Change-Id: I750a4d0e75e0b919fd00ddf21c0e7ce62d495f95