Fix debug tox environment
Without this patch, inserting a breakpoint causes the debug tox
environment to hang for a long time until the testenv timeout is
reached. This patch modifies the testenv to use similar stdout/stderr
settings that we use in keystoneclient and keystoneauth, which seems to
fix the issue, and removes other unnecessary settings.

Change-Id: I2f2f8f4738f43648a6bda067efe605db5807eaff