Run lower-constraints job on Xenial
The openstack-lower-constraints-jobs was updated to run in Ubuntu
Bionic, but the underlying version of openssl changed, causing tests
in keystonemiddleware to fail with the current version of
python-keystoneclient:

  https://review.openstack.org/#/c/406175/

Instead of bumping the version immediately, we can ensure the
lower-constraints job runs on Xenial for the time being, making it so
we can backport this fix to stable/stein. A subsequent patch will
update the job to use Bionic when we bump the minimum version of
python-keystoneclient.

Change-Id: I5a5ad8ad86df80755a304f70597578b7dfec2068