Comment html_static_path entry in docs conf.py
Currently with sphinx 2.2.0 the docs job is throwing a warning
that the html_static_path entry does not exist. We treat warnings
as errors so this causes the job to fail.

This change comments the html_static_path entry in conf.py since
the path currently does not exist so it appears to be unused.

Change-Id: Ib2c74f4f37855cec250d09b23c45b5b7fde44c8d