Make tox.ini tox 4.0.0 compatible/fix gate
* Removed skipsdist=True to make sure placement available in the virtual
  env. Without this, our entrypoints are not available.

* Removed basepython = python3 as we assume all developer switched to
  python3 in their env already

* Removed ignore_basepython_conflict = True as without the basepython
  definition generative targets now work without conflict

See [1] for a similar change made to placement.

It is also necessary to fix issues with the gate. For reasons that I
have yet to grok, a mock of 'requests.request' used in some test is no
longer functioning as expected. My guess is that something is now
importing requests before us and interfering with the mock but never
mind - we can easily bypass the issue by mocking 'requests.post'
instead.

Changes:
  keystonemiddleware/tests/unit/test_ec2_token_middleware.py
  test-requirements.txt

NOTE(stephenfin): It is necessary to revert the test changes since they
don't apply to this branch. It is also necessary to cap bandit since we
can't fix the timeout errors it is warning about without a feature
backport.

[1] https://review.opendev.org/c/openstack/placement/+/868418/

Change-Id: I3b8263afbf0ccee88ceaac2040d5ad274f22d74a
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
(cherry picked from commit 626df3a5e9e9d4fbea2649a9cfa0048083bdb88b)