cache: change log level for handled token corruption
When encountering a corrupted token from the cache, the previous
behavior was to log the event as an error. However, this situation is
expected and correctly handled in some scenarios. This commit changes
the log level to 'info' to keep operators aware without treating it as
an environment error that affects production.

Related-to: #2023015
Change-Id: I5512fca4eb994a72ef79c2b823f450e2f05a7bcf
Signed-off-by: Sahid Orentino Ferdjaoui <sahid.ferdjaoui@industrialdiscipline.com>