Wipe context before returning request response
This ensures that the next request in this thread will not reuse an old
context for a previous request.

That can happen in scenarios where the subsequent request fails auth
and no new context is generated to overwrite it.

As a result, requests can be logged with incorrect auth data from old
requests which is confusing for operators.

Related-Bug: #2130049
Change-Id: Icd07abe38a1201a488abee2bc13f8abf987e0672
Signed-off-by: Adam Oswick <adam@adamoswick.co.uk>